+++ R/plot_distribs.R
@@ -39,7 +39,8 @@ plot_distribs <- function(
   parname=c('GPP_daily','alpha','Pmax','ER_daily','K600_daily',
             'K600_daily_meanlog','lnK600_lnQ_intercept','lnK600_lnQ_slope','K600_lnQ_nodes',
             'K600_daily_sdlog','K600_daily_sigma',
-            'err_obs_iid_sigma','err_proc_acor_phi','err_proc_acor_sigma','err_proc_iid_sigma'), 
+            'err_obs_iid_sigma','err_proc_acor_phi','err_proc_acor_sigma','err_proc_iid_sigma',
+            'err_mult_GPP_sdlog'), 
   index=TRUE,
   style=c('dygraphs','ggplot2')) {
   
@@ -93,7 +94,8 @@ plot_distribs <- function(
     err_obs_iid_sigma='halfcauchy',
     err_proc_acor_phi='beta',
     err_proc_acor_sigma='halfcauchy',
-    err_proc_iid_sigma='halfcauchy'
+    err_proc_iid_sigma='halfcauchy',
+    err_mult_GPP_sdlog='halfnormal'
   )[parname]
   
   # create a data.frame illustrating the prior distribution